Deeplearning_project/ocr-from-scratch/api/requirements.txt
# OCR API — Python dependencies
# Install with: pip install -r requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9       # required by FastAPI for file uploads
pydantic>=2.7.0

# ML
torch>=2.3.0
torchvision>=0.18.0
Pillow>=10.0.0
